latest commit

+++ 11-terminate-instances.sh
@@ -1,14 +1,15 @@
 #!/bin/bash
 
+terminate_instances() {
+  ids=$(aws ec2 describe-instances \
+    --region us-east-1 \
+    --query "Reservations[].Instances[?InstanceType=='t3.micro'].InstanceId" \
+    --output text)
 
-terminate_instances(){
-
-instances=$(aws ec2 describe-instances --region us-east-1 \ --query "Reservations[*].Instances[*].[InstanceId,InstanceType]" \ --output text)
-
-ids=$(echo "$instances" | awk '$2=="t3.micro" {print $1}')
-
-aws ec2 terminate-instances --region us-east-1 --instance-ids $ids
-
+  if [ -n "$ids" ]; then
+    echo "Terminating instances: $ids"
+    aws ec2 terminate-instances --region us-east-1 --instance-ids $ids
+  else
+    echo "No t3.micro instances found."
+  fi
 }
-
-terminate_instances
